requirements.txt
serpapi~=0.1.5
python-dotenv~=1.0.1
pandas~=2.2.3
matplotlib~=3.9.2
pymongo~=4.10.1
mysql-connector-python
SQLAlchemy~=2.0.35
firebase-admin
ujson
flask~=3.0.3

numpy~=2.1.2
requests~=2.32.3
seaborn~=0.13.2